Adopt params to a 240Pu case similar to my data
Adopt parameters to 240Pu, with parameters similar to my data

I also added a file, 240PuPop_combEB06.dat, that is used in the ExFullRxn. It is a combination of some distriution calculated from TALYS for lower energies, and a "self" defined Jpop distribution for higher energies, ca 4-6MeV. In this case I wanted to see how it looks like if Jpop is given by EB06.