Fix #1 - Correct h2JIntrins
Fixing how the histogram of the underlying Js is created.

Previsouly a TF2 function was used to retrieve the spin distribution. This lead to problems, as it requested the distribution for non (half)integer values. Additionally, it could result in cumbersome results for E<E_critical.

Now the histogram is created by a loop over all continuum Ex bins and the spins that are to be plotted. The array adJInt is filled with the spin distribution at each excitation energy(bin). After normalization it is filled in g_h2JIntrins.

Finally, the errors are set to 0. By default root would set the Sumofw2 to sqrt(N). However, the spin distribution has no statistical uncertainties. (Rather systematic uncertainties on which distribution to use).